Use SHA-256 for generating SKI
