+++ cryptoutil/doc.go
@@ -1,0 +1,2 @@
+// package cryptoutil provides utilities for working with crypto types.
+package cryptoutil
